refactor sortable, formatted text editor into own reusable component
